# openapi.yaml
openapi: "3.0.1"
info:
  title: "Secret Hitler"
  description: "Web app companion to playing Secret Hitler"
  contact:
    name: "Sean Murphy"
    url: "www.murphysean.com"
    email: "murphysean84@gmail.com"
  license:
    name: "MIT"
    url: "https://github.com/payne8/secret-h/blob/master/LICENSE"
  version: "1.0.0"
servers:
  - url: "secrethitler.murphysean.com"
    description: "Cononical version of the application"
paths:
...
  /api/games:
    summary: "Games API"
    description: "Create and List games"
    get:
      tags: ["api"]
      summary: "List of games"
      responses:
        200:
          description: "A list of active games"
    post:
      tags: ["api"]
      summary: "Create a new game"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/game"
      responses:
        200:
          description: "The created game"
  /api/games/{gameId}:
    parameters:
      - name: "gameId"
        schema:
          type: string
          format: uuid
        in: "path"
        required: true
    get:
      tags: ["api"]
      responses:
        200:
          description: "A game object"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/game"
    put:
      tags: ["api"]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/game"
      responses:
        200:
          description: "The updated game object"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/game"
...
  /api/players:
    post:
      tags: ["api"]
      summary: "Create a new player"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/apiPlayer"
      responses:
        200:
          description: "The created player object"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apiPlayer"
        default:
          $ref: "#/components/responses/jsonError"
...
components:
  schemas:
    policy:
      type: string
      enum: ["liberal","fascist","masked"]
    party:
      type: string
      enum: ["liberal","fascist","masked"]
    role:
      type: string
      enum: ["liberal","fascist","hitler","masked"]
    gameState:
      type: string
      enum: ["","init","started","finished"]
    roundState:
      type: string
      enum: ["","voting","failed","legislating","executive_action","finished"]
    executiveAction:
      type: string
      enum: ["","investigate","peek","special_election","execute"]
    game:
      type: object
      properties:
        id:
          type: string
        secret:
          type: string
        eventId:
          type: number
        state:
          $ref: "#/components/schemas/gameState"
        draw:
          type: array
          items:
            $ref: "#/components/schemas/policy"
        discard:
          type: array
          items:
            $ref: "#/components/schemas/policy"
        liberal:
          type: number
          minimum: 0
          maximum: 5
        fascist:
          type: number
          minimum: 0
          maximum: 6
        electionTracker:
          type: number
          minimum: 0
          maximum: 3
        players:
          type: array
          items:
            $ref: "#/components/schemas/player"
        round:
          $ref: "#/components/schemas/round"
        nextPresidentId:
          type: string
        previousPresidentId:
          type: string
        previousChancellorId:
          type: string
        previousEnactedPolicy:
          type: string
        specialElectionRoundId:
          type: number
        specialElectionPresidentId:
          type: string
        winningParty:
          $ref: "#/components/schemas/party"
    round:
      type: object
      properties:
        id:
          type: number
        presidentId:
          type: string
        chancellorId:
          type: string
        state:
          $ref: "#/components/schemas/roundState"
        votes:
          type: array
          items:
            $ref: "#/components/schemas/vote"
        policies:
          type: array
          items:
            $ref: "#/components/schemas/policy"
        enactedPolicy:
          $ref: "#/components/schemas/policy"
        executiveAction:
          $ref: "#/components/schemas/executiveAction"
    vote:
      type: object
      properties:
        playerId:
          type: string
        vote:
          type: boolean
    player:
      type: object
      properties:
        id:
          type: string
        party:
          $ref: "#/components/schemas/party"
        role:
          $ref: "#/components/schemas/role"
        ready:
          type: boolean
        ack:
          type: boolean
        executedBy:
          type: string
        investigatedBy:
          type: string
        lastAction:
          type: string
          format: dateTime
        status:
          type: string
    apiPlayer:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        name:
          type: string
        thumbnailURL:
          type: string
          format: url
        password:
          type: string
          format: password
...
  responses:
    jsonError:
      description: "Something went wrong"
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: string
  securitySchemes:
    cookie:
      type: "apiKey"
      description: "Auth Cookie"
      name: "shsid"
      in: "cookie"
    header:
      type: "apiKey"
      description: "Bearer token"
      name: "Authorization"
      in: "header"
security:
  - authenticated_user: []
  - joined_user: []
  - administrator: []